requirements.txt
flask==3.0.0
anthropic>=0.40.0
python-dotenv==1.0.0
playwright==1.40.0
PyPDF2==3.0.1
python-docx==1.1.0
flask-session==0.5.0
werkzeug==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0